requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Argon2==0.3.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
pymongo==4.7.3
PyYAML==6.0.1
Werkzeug==3.0.3
